Form State Hook: useActionState This hook is used to manage and track the state of the form while interacting with a server action:
createComment.bind(null, articleId) binds the article ID to the action in advance. This avoids passing it again inside the form.

The second argument, { errors: {} }, sets the initial state of the form to an object with an empty errors field.

The hook returns three values:

formState: Contains the result or any errors returned from the server.

action: The bound server function that will be executed on form submission.

isPending: A boolean that is true while the server action is in progress.