Input Field:
The Input field is of type search and is used for user input.

It uses defaultValue={params.get("search") || ""} to pre-fill the input with the value from the URL query parameters (if available). If the search query parameter is not present, it will default to an empty string.

The input field has padding on the left (pl-10) to accommodate the search icon without overlapping the text.

The class w-48 gives the input field a fixed width, and focus-visible:ring-1 applies a ring effect when the field is focused.