Fetching Article Data The component is async because it needs to fetch the article data from a database. It performs the following:
Destructuring id from params: It extracts the id of the article from the params prop. Since params is a promise, it uses await to resolve the promise and access the id value.

Fetching Article Using Prisma:

The prisma.articles.findUnique() method is used to query the database for an article with the given id. The where clause specifies the condition that the article's id should match the provided id.

The include clause specifies that the article's author details (name, email, and image URL) should be fetched alongside the article data.