Configuring the Middleware (The config Object) The config object defines which routes the clerkMiddleware should be applied to and which routes should be excluded.
matcher: This is an array that defines the patterns of the routes (URLs) for which the middleware should run. It contains two parts:

First Part: A regular expression that matches any route except the ones that correspond to Next.js internals and static files. The regular expression is a negative lookahead ((?!...)) to exclude specific patterns.

Second Part: The routes that match /api or /trpc paths will always run the middleware, ensuring that API and trpc routes are always protected.